backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
email-validator
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
groq
httpx
